+++ src/components/Recommendations.jsx
@@ -1,24 +1,19 @@
 import React, { useEffect, useState } from 'react'
+import AliceCarousel from 'react-alice-carousel'
+import 'react-alice-carousel/lib/alice-carousel.css'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { Redirect } from 'react-router-dom'
-import {
-  Paper,
-  Typography,
-  Divider,
-  Grid,
-  Button
-} from '@material-ui/core'
+import { Button } from '@material-ui/core/'
 
 import { RecommendationsStyles } from '../style/muiStyles'
+import { getMovies } from '../actions/movies'
 import RecommendationDetail from './RecommendationDetail'
 
-import { getMovies } from '../actions/movies'
-
-const Recommendations = ({ movies, info, dispatch }, ...props) => {
+const RecommendationsNew = ({ movies, info, dispatch }, ...props) => {
   const classes = RecommendationsStyles(props)
-
   const [redirect, setRedirect] = useState()
+  const handleOnDragStart = e => e.preventDefault()
 
   useEffect(() => {
     dispatch(getMovies())
@@ -32,6 +27,18 @@ const Recommendations = ({ movies, info, dispatch }, ...props) => {
     setRedirect('see')
   }
 
+  const responsive = {
+    1024: {
+      items: 3
+    },
+    464: {
+      items: 2
+    },
+    0: {
+      items: 1
+    }
+  }
+
   const renderRedirect = () => {
     if (redirect === 'see') {
       return <Redirect push to={'/movies/list'} />
@@ -41,28 +48,37 @@ const Recommendations = ({ movies, info, dispatch }, ...props) => {
   }
 
   return (
-    movies && (
-      <Paper>
-        {redirect && renderRedirect()}
-        <Typography variant="h6" component="h2" gutterBottom>
-          OUR TOP PICKS
-        </Typography>
-        <Divider />
-        <Grid container direction="row" justify="center" alignItems="center">
-          {movies
-            .filter(movie => movie.recommended)
-            .map(movie => (
-              <RecommendationDetail
-                movie={movie}
-                id={movie.id}
-                key={movie.id}
-                onClick={handleClick}
-                image={`https://image.tmdb.org/t/p/w200${movie.image}`}
-                title={movie.title} />
-            ))}
-        </Grid>
-        <Button variant="outlined" color="primary" className={classes.button} onClick={handleSeeAll}>SEE ALL MOVIES</Button>
-      </Paper>
+    !info.pending && (
+      <>
+        <div classes={classes.carousel}>
+          {redirect && renderRedirect()}
+          <AliceCarousel
+            responsive={responsive}
+            autoPlayInterval={3300}
+            duration={500}
+            autoPlayDirection="ltr"
+            autoPlay={true}
+            fadeOutAnimation={false}
+            mouseDragEnabled={true}
+            dotsDisabled={true}
+            buttonsDisabled={true}>
+            {movies
+              .filter(movie => movie.recommended)
+              .map(movie => (
+                <RecommendationDetail
+                  onDragStart={handleOnDragStart}
+                  movieTests={movie.movieTests}
+                  id={movie.id}
+                  key={movie.id}
+                  onClick={handleClick}
+                  image={`https://image.tmdb.org/t/p/w500${movie.image}`}
+                  title={movie.title}
+                  year={movie.releaseDate} />
+              ))}
+          </AliceCarousel>
+          <Button variant="outlined" color="primary" className={classes.button} onClick={handleSeeAll}>SEE ALL MOVIES</Button>
+        </div>
+      </>
     )
   )
 }
@@ -74,10 +90,10 @@ function mapStateToProps ({ movies, info }) {
   }
 }
 
-export default connect(mapStateToProps)(Recommendations)
-
-Recommendations.propTypes = {
+RecommendationsNew.propTypes = {
   dispatch: PropTypes.func,
   movies: PropTypes.array,
   info: PropTypes.object
 }
+
+export default connect(mapStateToProps)(RecommendationsNew)
